requirements.txt
requests
praw
pandas
kafka-python
pyspark
dotenv